Revert "Accidentally invented a module system"
This reverts commit 973b8b11f8ee7cf8dc6bb5f70ab5b084acb59035.
