+++ css/custom.css
@@ -211,16 +211,3 @@ body {
 .gallery-page.active-page {
     display: block;
 }
-
-.notification-enable {
-    width: 100%;
-}
-
-.slow .toggle-group {
-    transition: left 0.7s;
-    -webkit-transition: left 0.7s;
-}
-
-#notification-state {
-    margin-top: 5px;
-}
