+++ js/action.js
@@ -1,7 +1,6 @@
 $( document ).ready(function() {
     setLanguage();
     setBackground();
-    setNotificationsState();
     setSoundpack();
 
     /* ### Volume Settings ### */
@@ -89,8 +88,4 @@ $( document ).ready(function() {
     $('#previous-page').click(function(){
         previousGalleryPage();
     });
-
-    $('#enable-notifications').change(function() {
-      setStorage('get-notifications', $(this).prop('checked'));
-    })
 });
