--- modules/notifications/settings.php
@@ -1,6 +1,0 @@
-<div class="settings-col">
-    <div class="col-md-6"><?php $this->oTranslator->TranslateString("notification-settings") ?></div>
-    <div class="col-md-6">
-        <button type="button" class="btn btn-default" data-toggle="modal" data-target="#sound-gallery"><?php $this->oTranslator->TranslateString("notification-settings-btn") ?></button>
-    </div>
-</div>
